requirements.txt
anthropic>=0.49.0
beautifulsoup4>=4.13.3
chromadb>=0.6.3
faiss-cpu>=1.10.0
html2text>=2024.2.26
httpx>=0.28.1
langchain-community>=0.3.20
langchain>=0.3.22
mysql-connector-python>=9.2.0
numpy>=2.2.4
ollama>=0.4.7
openai>=1.68.2
pandas>=2.2.3
pdf2image>=1.17.0
plotly>=6.0.1
psycopg2-binary>=2.9.10
pydantic>=2.10.6
pypdf>=5.4.0
python-dotenv>=1.1.0
qdrant-client>=1.13.3
requests>=2.32.3
selenium>=4.30.0
spacy>=3.8.5
streamlit>=1.44.0
tiktoken>=0.9.0
trafilatura>=2.0.0
webdriver-manager>=4.0.2
langchain-core>=0.3.49
pdfplumber>=0.11.6
langchain-openai>=0.3.12
langchain-qdrant>=0.2.0
apify-client>=1.9.3
